Refactor subsystem code
