+++ conan/tools/microsoft/__init__.py
@@ -2,5 +2,5 @@ from conan.tools.microsoft.toolchain import MSBuildToolchain
 from conan.tools.microsoft.msbuild import MSBuild
 from conan.tools.microsoft.msbuilddeps import MSBuildDeps
 from conan.tools.microsoft.visual import msvc_runtime_flag, VCVars, is_msvc, is_msvc_static_runtime
-from conan.tools.microsoft.subsystems import subsystem_path
 from conan.tools.microsoft.layout import vs_layout
+from conan.tools.microsoft.subsystems import unix_path
